# api/docs/swagger.yaml
definitions:
  models.Basket:
    properties:
      ctreate_at:
        type: string
      id:
        type: string
      price:
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  models.BasketResponse:
    properties:
      baskets:
        items:
          $ref: '#/definitions/models.Basket'
        type: array
      count:
        type: integer
    type: object
  models.Branch:
    properties:
      address:
        type: string
      ctreate_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.CreateBasket:
    properties:
      ctreate_at:
        type: string
      price:
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  models.CreateBranch:
    properties:
      address:
        type: string
      ctreate_at:
        type: string
      name:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateBasket:
    properties:
      ctreate_at:
        type: string
      id:
        type: string
      price:
        type: integer
      product_id:
        type: string
      quantity:
        type: integer
      sale_id:
        type: string
    type: object
  models.UpdateBranch:
    properties:
      address:
        type: string
      ctreate_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /basket:
    post:
      consumes:
      - application/json
      description: create a new basket
      parameters:
      - description: basket
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateBasket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new basket
      tags:
      - basket
  /basket/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Basket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete basket by ID
      tags:
      - basket
    get:
      consumes:
      - application/json
      description: get basket by ID
      parameters:
      - description: basket
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Gets basket
      tags:
      - basket
    put:
      consumes:
      - application/json
      description: update basket
      parameters:
      - description: basket
        in: path
        name: id
        required: true
        type: string
      - description: basket
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBasket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Basket'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update basket
      tags:
      - basket
  /baskets:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BasketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get a list of baskets
      tags:
      - basket
  /branch:
    post:
      consumes:
      - application/json
      description: create a new branch
      parameters:
      - description: branch
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateBranch'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Branch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new branch
      tags:
      - branch
  /branch/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete branch by ID
      tags:
      - branch
    get:
      consumes:
      - application/json
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Branch'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get branch by ID
      tags:
      - branch
    put:
      consumes:
      - application/json
      description: update branch
      parameters:
      - description: branch
        in: path
        name: id
        required: true
        type: string
      - description: branch
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Branch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update branch
      tags:
      - branch
  /branchs:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Branch'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a list of branches
      tags:
      - branch
...
swagger: "2.0"